python_server/requirements.txt
paho-mqtt
SQLAlchemy